refactor: separated draw from universe
